[TASK] Get overwriteDemand variable from session if it is not set in PerformanceController and EventController
For <f:widget.link> is the attribute addQueryStringMethod="POST,GET" deprecated. That means that the pagination does not work any more because it doesn't use the current overwriteDemand parameters. To fix this, it will be stored in the overwriteDemand into the session and it will be removed only when the request paremeter "resetDemand" is set.